Refactor backends module to avoid importing deps on models import
